requirements.txt
pyttsx3==2.90
simple_colors==0.1.5
SpeechRecognition==3.8.1
requests==2.26.0
python-dotenv==0.19.0
